requirements.txt
python-dotenv==1.0.0
groq==0.4.0
openai==1.13.3
anthropic==0.18.1
tiktoken==0.6.0
prompt_toolkit==3.0.43